corrected integral division error and removed the confusing unused "inplacemutation" class (which did not correspond to the "in place" config option)
git-svn-id: file:///srv/local/data/cleanRepo/trunk@17478 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/thibautd/src/main/java/playground/thibautd/jointtripsoptimizer/replanning/modules/geneticoperators/JointPlanOptimizerJGAPMutation.java
@@ -303,7 +303,7 @@ public class JointPlanOptimizerJGAPMutation implements GeneticOperator {
 
 	private final double delta(final int t,final double y) {
 		double r = this.randomGenerator.nextDouble();
-		double exponant = Math.pow((1d - t/this.NUM_ITER), this.NON_UNIFORMITY_PARAM);
+		double exponant = Math.pow((1d - ((double) t)/this.NUM_ITER), this.NON_UNIFORMITY_PARAM);
 
 		return (y * (1 - Math.pow(r, exponant)));
 	}
